feat: implement subtasks management, skill catalog APIs, HITL AI feedback verification, and ML dataset export enhancements - Add Subtask entity, database migration script, and CRUD / status-toggle APIs under /api/v1/tasks/{taskId}/sub-tasks - Add Skills and ResourceSkills controllers and services for managing skill catalog, member proficiency levels, and categories - Add Human-In-The-Loop (HITL) verification and feedback API under /api/v1/projects/{projectId}/ai-insights/{logId}/feedback - Update Python AI assessment DTOs to support semantic skill scoring, matched skills breakdown, and marginal match detection - Enhance MLExportService and AILog schema to capture HITL verification labels, feedback ratings, and enriched feature metrics - Add SQL migration scripts for subtasks and HITL verification columns, and configure Dependency Injection in Program.cs

+++ src/AllocServer/Models/ProjectRiskFeature.cs
@@ -10,7 +10,7 @@ namespace AllocServer.Models
         public int Total_Tasks { get; set; }
         public int Team_Size { get; set; }
         public double Avg_Team_Skill_Level { get; set; }
-        public double Raw_Complexity_Score { get; set; }
+        public decimal Raw_Complexity_Score { get; set; }
         public decimal Budget_Utilization_Rate { get; set; }
         public double Overall_Risk_Score { get; set; }
     }
